pyproject.toml
[project]
name = "horen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.4",
    "black>=25.1.0",
    "fastapi>=0.116.1",
    "icecream>=2.1.7",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]

[dependency-groups]
dev = [
    "ruff>=0.12.9",
]
